pyproject.toml
[project]
name = "crehana-proyect"
version = "0.1.0"
description = ""
authors = [
    {name = "orlando salas"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt<4.0.0",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "binary (>=1.0.2,<2.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
pytest-cov = "^6.2.1"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
  | env
  | venv
  | \.venv
)/
'''